Update README.md
The aim of this project is to forecast the demand for food delivery services for a given area. The dataset used for this project contains information on the orders made on a food delivery platform over a period of several months. The dataset includes information such as the date and time of the order, the location of the restaurant, and the type of cuisine.